Extract iso language list to new file
